backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-dotenv
pymongo
python-jose
passlib
bcrypt>=4.0.0
jwt